build: add backend-only server
