+++ src/sequelize.ts
@@ -1,15 +1,17 @@
+import path from 'path'
 import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
 
 // DB_ENV is used soley to set the database config to "ci" for running
 // tests on CI. This is because DB connection config is different between
 // CI and local test envs. See db/config.json
 const env = process.env.DB_ENV || process.env.NODE_ENV || 'development'
-const config = require(__dirname + '/../db/config.json')[env]
+const config = require(path.join(process.cwd(), 'db', 'config.json'))[env]
+console.debug(`sequelize config loaded`, config)
 
 export let sequelize: Sequelize
 
 const COMMON_CONFIG: Partial<SequelizeOptions> = {
-  models: [__dirname + '/models'],
+  models: [path.join(process.cwd(), 'dist', 'src', 'models')],
   dialect: 'postgres',
   protocol: 'postgres',
 }
